Clean up and bundle commands code
