+++ trustpoint/pki/management/commands/test_long_chain_p12.py
@@ -3,109 +3,18 @@
 
 from __future__ import annotations
 
-
-from django.core.management import BaseCommand
-
-from cryptography import x509
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.serialization import pkcs12, BestAvailableEncryption
-from cryptography.hazmat.primitives.asymmetric import rsa, ec
-from cryptography.x509.oid import NameOID
-import datetime
-
 from pathlib import Path
 
-class Command(BaseCommand):
+from cryptography.hazmat.primitives.serialization import BestAvailableEncryption, pkcs12
+from pki.models import CertificateModel
+
+from .base_commands import Command
+
+
+class Command(Command):
     """Django management command for adding issuing CA test data."""
 
     help = 'Removes all migrations, deletes db and runs makemigrations and migrate afterwards.'
-
-    @staticmethod
-    def create_root_ca(cn: str) -> tuple[x509.Certificate, rsa.RSAPrivateKey]:
-        one_day = datetime.timedelta(365, 0, 0)
-        private_key = rsa.generate_private_key(
-            public_exponent=65537,
-            key_size=2048,
-        )
-        public_key = private_key.public_key()
-        builder = x509.CertificateBuilder()
-        builder = builder.subject_name(x509.Name([
-            x509.NameAttribute(NameOID.COMMON_NAME, cn),
-        ]))
-        builder = builder.issuer_name(x509.Name([
-            x509.NameAttribute(NameOID.COMMON_NAME, cn),
-        ]))
-        builder = builder.not_valid_before(datetime.datetime.today() - one_day)
-        builder = builder.not_valid_after(datetime.datetime.today() + (one_day * 30))
-        builder = builder.serial_number(x509.random_serial_number())
-        builder = builder.public_key(public_key)
-        builder = builder.add_extension(
-            x509.BasicConstraints(ca=True, path_length=None), critical=True,
-        )
-        certificate = builder.sign(
-            private_key=private_key, algorithm=hashes.SHA256(),
-        )
-        return certificate, private_key
-
-    @staticmethod
-    def create_issuing_ca(
-            issuer_private_key: rsa.RSAPrivateKey,
-            issuer_cn: str, subject_cn,
-            private_key: None | rsa.RSAPrivateKey = None) -> tuple[x509.Certificate, rsa.RSAPrivateKey]:
-        one_day = datetime.timedelta(365, 0, 0)
-        if private_key is None:
-            private_key = rsa.generate_private_key(
-                public_exponent=65537,
-                key_size=2048,
-            )
-        public_key = private_key.public_key()
-        builder = x509.CertificateBuilder()
-        builder = builder.subject_name(x509.Name([
-            x509.NameAttribute(NameOID.COMMON_NAME, subject_cn),
-        ]))
-        builder = builder.issuer_name(x509.Name([
-            x509.NameAttribute(NameOID.COMMON_NAME, issuer_cn),
-        ]))
-        builder = builder.not_valid_before(datetime.datetime.today() - one_day)
-        builder = builder.not_valid_after(datetime.datetime.today() + (one_day * 30))
-        builder = builder.serial_number(x509.random_serial_number())
-        builder = builder.public_key(public_key)
-        builder = builder.add_extension(
-            x509.BasicConstraints(ca=True, path_length=None), critical=True,
-        )
-        certificate = builder.sign(
-            private_key=issuer_private_key, algorithm=hashes.SHA256(),
-        )
-        return certificate, private_key
-
-    @staticmethod
-    def create_ee(
-            issuer_private_key: rsa.RSAPrivateKey,
-            issuer_cn: str, subject_cn) -> tuple[x509.Certificate, rsa.RSAPrivateKey]:
-        one_day = datetime.timedelta(365, 0, 0)
-        private_key = rsa.generate_private_key(
-            public_exponent=65537,
-            key_size=2048,
-        )
-        public_key = private_key.public_key()
-        builder = x509.CertificateBuilder()
-        builder = builder.subject_name(x509.Name([
-            x509.NameAttribute(NameOID.COMMON_NAME, subject_cn),
-        ]))
-        builder = builder.issuer_name(x509.Name([
-            x509.NameAttribute(NameOID.COMMON_NAME, issuer_cn),
-        ]))
-        builder = builder.not_valid_before(datetime.datetime.today() - one_day)
-        builder = builder.not_valid_after(datetime.datetime.today() + (one_day * 30))
-        builder = builder.serial_number(x509.random_serial_number())
-        builder = builder.public_key(public_key)
-        builder = builder.add_extension(
-            x509.BasicConstraints(ca=False, path_length=None), critical=True,
-        )
-        certificate = builder.sign(
-            private_key=issuer_private_key, algorithm=hashes.SHA256(),
-        )
-        return certificate, private_key
 
     def handle(self, *args, **kwargs) -> None:
 
@@ -122,7 +31,7 @@ class Command(BaseCommand):
         issuing_5, issuing_5_key = self.create_issuing_ca(
             issuing_4_key, 'Intermediate CA D', 'Issuing CA')
 
-        # ee_1, ee_key = self.create_ee(issuing_5_key, 'Issuing CA', 'EE A1')
+        ee_1, ee_key = self.create_ee(issuing_5_key, 'Issuing CA', 'EE A1')
 
         p12 = pkcs12.serialize_key_and_certificates(
             b'my p12',
@@ -134,12 +43,12 @@ class Command(BaseCommand):
 
         with open(Path(__file__).parent.parent.parent / 'p12.p12', 'wb') as f:
             f.write(p12)
-        # CertificateModel.save_certificate(root_1)
-        # CertificateModel.save_certificate(issuing_1)
-        # CertificateModel.save_certificate(issuing_2)
-        # CertificateModel.save_certificate(issuing_3)
-        # CertificateModel.save_certificate(issuing_4)
-        # CertificateModel.save_certificate(issuing_5)
-        # CertificateModel.save_certificate(ee_1)
+        CertificateModel.save_certificate(root_1)
+        CertificateModel.save_certificate(issuing_1)
+        CertificateModel.save_certificate(issuing_2)
+        CertificateModel.save_certificate(issuing_3)
+        CertificateModel.save_certificate(issuing_4)
+        CertificateModel.save_certificate(issuing_5)
+        CertificateModel.save_certificate(ee_1)
