+++ trustpoint/pki/management/commands/test_ts.py
@@ -3,18 +3,18 @@
 
 from __future__ import annotations
 
+from itertools import product
 from pathlib import Path
+
 from cryptography import x509
 from cryptography.hazmat.primitives import serialization
-from itertools import product
-
-from django.core.management import BaseCommand
 from pki.models import CertificateModel
 
 from . import Algorithm
+from .base_commands import Command
 
 
-class Command(BaseCommand):
+class Command(Command):
     """Django management command for adding issuing CA test data."""
 
     help = 'Just some stuff for manual testing.'
